moved to extract_configuration_data and recursive_extract_configuration_data to their own file.
